Implementadas las colisiones y que se muestre la pantalla de Game Over

+++ src/MainGameState.hpp
@@ -3,12 +3,12 @@
 #include <deque>
 
 extern "C" {
-    #include "raylib.h"
+    #include "../vendor/include/raylib.h"
 }
 
 struct Bird
 {
-  int x, y;
+  float x, y;
   float vy;
 };
